refactor: update role handling and permissions logic, update types, update buttons
