+++ ui/lib/permissions.ts
@@ -1,0 +1,42 @@
+/**
+ * Get the permissions for a user role
+ * @param attributes - The attributes of the user role
+ * @returns The permissions for the user role
+ */
+export const getRolePermissions = (attributes: any) => {
+  const permissions = [
+    {
+      key: "manage_users",
+      label: "Manage Users",
+      enabled: attributes.manage_users,
+    },
+    {
+      key: "manage_account",
+      label: "Manage Account",
+      enabled: attributes.manage_account,
+    },
+    {
+      key: "manage_providers",
+      label: "Manage Providers",
+      enabled: attributes.manage_providers,
+    },
+    {
+      key: "manage_scans",
+      label: "Manage Scans",
+      enabled: attributes.manage_scans,
+    },
+
+    {
+      key: "manage_integrations",
+      label: "Manage Integrations",
+      enabled: attributes.manage_integrations,
+    },
+    {
+      key: "unlimited_visibility",
+      label: "Unlimited Visibility",
+      enabled: attributes.unlimited_visibility,
+    },
+  ];
+
+  return permissions;
+};
